+++ packages/eslint-plugin/src/rules/explicit-member-accessibility.ts
@@ -1,6 +1,8 @@
 import {
   AST_NODE_TYPES,
   TSESTree,
+  AST_TOKEN_TYPES,
+  TSESLint,
 } from '@typescript-eslint/experimental-utils';
 import * as util from '../util';
 
@@ -38,6 +40,7 @@ export default util.createRule<Options, MessageIds>({
       // too opinionated to be recommended
       recommended: false,
     },
+    fixable: 'code',
     messages: {
       missingAccessibility:
         'Missing accessibility modifier on {{type}} {{name}}.',
@@ -91,6 +94,7 @@ export default util.createRule<Options, MessageIds>({
       nodeType: string,
       node: TSESTree.Node,
       nodeName: string,
+      fix: TSESLint.ReportFixFunction | null = null,
     ): void {
       context.report({
         node: node,
@@ -99,6 +103,7 @@ export default util.createRule<Options, MessageIds>({
           type: nodeType,
           name: nodeName,
         },
+        fix: fix,
       });
     }
 
@@ -140,6 +145,7 @@ export default util.createRule<Options, MessageIds>({
           nodeType,
           methodDefinition,
           methodName,
+          getUnwantedPublicAccessibilityFixer(methodDefinition),
         );
       } else if (check === 'explicit' && !methodDefinition.accessibility) {
         reportIssue(
@@ -149,6 +155,47 @@ export default util.createRule<Options, MessageIds>({
           methodName,
         );
       }
+    }
+
+    /**
+     * Creates a fixer that removes a "public" keyword with following spaces
+     */
+    function getUnwantedPublicAccessibilityFixer(
+      node:
+        | TSESTree.MethodDefinition
+        | TSESTree.ClassProperty
+        | TSESTree.TSParameterProperty,
+    ): TSESLint.ReportFixFunction {
+      return function(fixer: TSESLint.RuleFixer): TSESLint.RuleFix {
+        const tokens = sourceCode.getTokens(node);
+        let rangeToRemove: TSESLint.AST.Range;
+        for (let i = 0; i < tokens.length; i++) {
+          const token = tokens[i];
+          if (
+            token.type === AST_TOKEN_TYPES.Keyword &&
+            token.value === 'public'
+          ) {
+            const commensAfterPublicKeyword = sourceCode.getCommentsAfter(
+              token,
+            );
+            if (commensAfterPublicKeyword.length) {
+              // public /* Hi there! */ static foo()
+              // ^^^^^^^
+              rangeToRemove = [
+                token.range[0],
+                commensAfterPublicKeyword[0].range[0],
+              ];
+              break;
+            } else {
+              // public static foo()
+              // ^^^^^^^
+              rangeToRemove = [token.range[0], tokens[i + 1].range[0]];
+              break;
+            }
+          }
+        }
+        return fixer.removeRange(rangeToRemove!);
+      };
     }
 
     /**
@@ -170,6 +217,7 @@ export default util.createRule<Options, MessageIds>({
           nodeType,
           classProperty,
           propertyName,
+          getUnwantedPublicAccessibilityFixer(classProperty),
         );
       } else if (propCheck === 'explicit' && !classProperty.accessibility) {
         reportIssue(
@@ -217,6 +265,7 @@ export default util.createRule<Options, MessageIds>({
               nodeType,
               node,
               nodeName,
+              getUnwantedPublicAccessibilityFixer(node),
             );
           }
           break;