+++ packages/eslint-plugin/src/rules/switch-exhaustiveness-check.ts
@@ -2,11 +2,12 @@ import { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
 import * as ts from 'typescript';
 import {
   createRule,
-  getParserServices,
   getConstrainedTypeAtLocation,
+  getParserServices,
+  isClosingBraceToken,
+  isOpeningBraceToken,
 } from '../util';
 import { isTypeFlagSet, unionTypeParts } from 'tsutils';
-import { isClosingBraceToken, isOpeningBraceToken } from 'eslint-utils';
 
 export default createRule({
   name: 'switch-exhaustiveness-check',