+++ packages/typescript-estree/src/ts-estree/ts-estree.ts
@@ -868,6 +868,8 @@ export interface EmptyStatement extends BaseNode {
 export interface ExportAllDeclaration extends BaseNode {
   type: AST_NODE_TYPES.ExportAllDeclaration;
   source: Expression | null;
+  exportKind: 'type' | 'value';
+  exported: Identifier | null;
 }
 
 export interface ExportDefaultDeclaration extends BaseNode {
@@ -880,6 +882,7 @@ export interface ExportNamedDeclaration extends BaseNode {
   declaration: ExportDeclaration | null;
   specifiers: ExportSpecifier[];
   source: Expression | null;
+  exportKind: 'type' | 'value';
 }
 
 export interface ExportSpecifier extends BaseNode {
@@ -949,6 +952,7 @@ export interface ImportDeclaration extends BaseNode {
   type: AST_NODE_TYPES.ImportDeclaration;
   source: Literal;
   specifiers: ImportClause[];
+  importKind: 'type' | 'value';
 }
 
 export interface ImportDefaultSpecifier extends BaseNode {